fix unit tests

+++ python_modules/libraries/dagster-duckdb/dagster_duckdb/io_manager.py
@@ -119,7 +119,7 @@ class DuckDbClient(DbClient):
     def get_select_statement(table_slice: TableSlice) -> str:
         col_str = ", ".join(table_slice.columns) if table_slice.columns else "*"
 
-        if len(table_slice.partition) > 0:
+        if table_slice.partition and len(table_slice.partition) > 0:
             query = f"SELECT {col_str} FROM {table_slice.schema}.{table_slice.table} WHERE\n"
             for i in range(len(table_slice.partition)):
                 part = table_slice.partition[i]
@@ -152,8 +152,8 @@ def _get_cleanup_statement(table_slice: TableSlice) -> str:
     Returns a SQL statement that deletes data in the given table to make way for the output data
     being written.
     """
-    if len(table_slice.partition) > 0:
-        query = f"DELETE FROM {table_slice.schema}.{table_slice.table} WHERE \n"
+    if table_slice.partition and len(table_slice.partition) > 0:
+        query = f"DELETE FROM {table_slice.schema}.{table_slice.table} WHERE\n"
         for i in range(len(table_slice.partition)):
             part = table_slice.partition[i]
             partition_where = (
